Split loader into multiple files
